Moved duplicated code up in hierarchy
